requirements.txt
# pypi.org/project/wordle-utils
wordle-utils==0.3.0
